requirements.txt
requests==2.31.0
pandas==2.2.3
numpy==2.2.0
plotly==5.18.0
streamlit==1.41.0
psycopg2-binary==2.9.11
boto3==1.34.0
apscheduler==3.10.4
python-dotenv==1.0.1
scikit-learn==1.5.2